fix: improve structure
