+++ frontend/src/features/chat/index.ts
@@ -1,8 +1,3 @@
 export { usePushChatQueue } from "./lib";
-export {
-  GomokuBoard,
-  GomokuChatCard,
-  GomokuInfoCard,
-  GomokuResultCard,
-} from "./ui";
+export { GomokuChatCard } from "./ui";
 export { chatQueueState } from "./model";
