+++ frontend/src/features/chat/ui/index.ts
@@ -1,4 +1,1 @@
-export { default as GomokuBoard } from "./GomokuBoard";
 export { default as GomokuChatCard } from "./GomokuChatCard";
-export { default as GomokuInfoCard } from "./GomokuInfoCard";
-export { default as GomokuResultCard } from "./GomokuResultCard";
