+++ frontend/src/features/gameplay/index.ts
@@ -1,0 +1,2 @@
+export { useGameplay } from "./lib";
+export { GomokuTimer, GomokuResult } from "./ui";
