+++ frontend/src/features/gameplay/lib/index.ts
@@ -1,0 +1,1 @@
+export { default as useGameplay } from "./useGameplay";
