+++ frontend/src/features/gameplay/ui/GomokuProgressBar.tsx
