+++ frontend/src/features/gameplay/ui/GomokuTimer.tsx
@@ -3,86 +3,21 @@ import { logo, vs, whiteStone, blackStone } from "@/shared/resource/images";
 import arrow from "./arrow.svg";
 
 import { colorStyles, textStyles } from "@/shared/ui";
-import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
-import { useEffect, useState } from "react";
-import useInterval from "use-interval";
-import { timer, move } from "@/shared/resource/audios";
+import { useRecoilValue } from "recoil";
+
 import { useTranslation } from "react-i18next";
 import { userState } from "@/entities/auth";
-import {
-  useMoveStone,
-  gomokuNowPlayerState,
-  gomokuResultState,
-  gomokuState,
-  gomokuTurnState,
-  gomokuVoteState,
-  str2numCoord,
-} from "@/entities/game";
+import { gomokuNowPlayerState, gomokuState } from "@/entities/game";
 import GomokuProgressBar from "./GomokuProgressBar";
+import { useGameplay } from "@/features/gameplay";
 
-const timerSound = new Audio(timer);
-const moveSound = new Audio(move);
-
-function GomokuInfoCard() {
+function GomokuTimer() {
   const { t } = useTranslation();
   const user = useRecoilValue(userState);
   const setting = useRecoilValue(gomokuState);
-  const turn = useRecoilValue(gomokuTurnState);
-  const [nowPlayer, setNowPlayer] = useRecoilState(gomokuNowPlayerState);
-  const [time, setTime] = useState(-1);
-  const [vote, setVote] = useRecoilState(gomokuVoteState);
-  const setResult = useSetRecoilState(gomokuResultState);
+  const nowPlayer = useRecoilValue(gomokuNowPlayerState);
 
-  const moveStone = useMoveStone();
-
-  useInterval(() => {
-    if (time === 6 && nowPlayer === 1) {
-      timerSound.play();
-    } else if (time > 6 && timerSound.played) {
-      timerSound.pause();
-      timerSound.currentTime = 0;
-    }
-    if (0 < time) {
-      setTime((ot) => ot - 1);
-    } else if (time === 0) {
-      if (nowPlayer === 1) {
-        setResult(2);
-      } else {
-        if (vote.total) {
-          let maxKey = "";
-          let maxValue = -1;
-
-          for (const [key, value] of vote.count) {
-            if (value > maxValue) {
-              maxKey = key;
-              maxValue = value;
-            }
-          }
-          const [i, j] = str2numCoord(maxKey);
-          moveStone(i, j, setting.viewerColor);
-        } else {
-          setResult(1);
-        }
-      }
-    }
-  }, 1000);
-
-  useEffect(() => {
-    if (setting.streamerColor === turn) {
-      setTime(setting.streamerTime);
-      setNowPlayer(1);
-    } else {
-      setTime(setting.viewerTime);
-      setNowPlayer(2);
-    }
-    moveSound.play();
-    setVote({ count: new Map(), total: 0 });
-
-    return () => {
-      timerSound.pause();
-      timerSound.currentTime = 0;
-    };
-  }, [turn]);
+  const { time } = useGameplay();
 
   return (
     <StyledGomokuInfoCard>
@@ -185,7 +120,7 @@ function GomokuInfoCard() {
   }
 }
 
-export default GomokuInfoCard;
+export default GomokuTimer;
 
 const StyledGomokuInfoCard = styled.section`
   padding: 8px;
