+++ frontend/src/features/gameplay/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as GomokuTimer } from "./GomokuTimer";
+export { default as GomokuResult } from "./GomokuResult";
