+++ frontend/src/widgets/gomoku-board/index.ts
@@ -1,0 +1,1 @@
+export { GomokuBoard } from "./ui";
