+++ frontend/src/widgets/gomoku-board/ui/GomokuBoard.tsx
