+++ frontend/src/widgets/gomoku-board/ui/GomokuBoardSquare.tsx
