+++ frontend/src/widgets/gomoku-board/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as GomokuBoard } from "./GomokuBoard";
